+++ src/main/java/edu/ifbasaj/pweb/cadastro_de_funcionarios/departamento/service/impl/DepartamentoServiceImpl.java
@@ -5,59 +5,45 @@ import java.util.ArrayList;
 import java.util.Optional;
 import java.util.UUID;
 
-import org.springframework.context.annotation.Lazy;
+
 import org.springframework.stereotype.Service;
 
 import edu.ifbasaj.pweb.cadastro_de_funcionarios.departamento.mapper.DepartamentoMapper;
 import edu.ifbasaj.pweb.cadastro_de_funcionarios.departamento.model.dto.DepartamentoDTO;
 import edu.ifbasaj.pweb.cadastro_de_funcionarios.departamento.repository.DepartamentoRepository;
 import edu.ifbasaj.pweb.cadastro_de_funcionarios.departamento.service.DepartamentoService;
-import edu.ifbasaj.pweb.cadastro_de_funcionarios.exceptions.CampoDisponivelVazioException;
+import edu.ifbasaj.pweb.cadastro_de_funcionarios.exceptions.CampoVazioException;
 import edu.ifbasaj.pweb.cadastro_de_funcionarios.exceptions.GerenteJaAssociadoException;
-import edu.ifbasaj.pweb.cadastro_de_funcionarios.exceptions.RemocaoDepartamentoGerenteException;
-import edu.ifbasaj.pweb.cadastro_de_funcionarios.funcionario.model.dto.FuncionarioDTO;
-import edu.ifbasaj.pweb.cadastro_de_funcionarios.funcionario.service.FuncionarioService;
+import edu.ifbasaj.pweb.cadastro_de_funcionarios.utils.ServiceUtils;
+import edu.ifbasaj.pweb.cadastro_de_funcionarios.exceptions.EntidadeAssociadaException;
 import jakarta.persistence.EntityNotFoundException;
+import lombok.RequiredArgsConstructor;
 
 @Service
+@RequiredArgsConstructor
 public class DepartamentoServiceImpl implements DepartamentoService {
     
     private final DepartamentoRepository repository;
     private final DepartamentoMapper mapper;
-    private final FuncionarioService funcionarioService;
-
-    public DepartamentoServiceImpl(DepartamentoRepository repository, DepartamentoMapper mapper,
-        @Lazy FuncionarioService funcionarioService){
-        
-        this.repository = repository;
-        this.mapper = mapper;
-        this.funcionarioService = funcionarioService;
-    }
+    private final ServiceUtils utils;
     
     @Override
     public Optional<DepartamentoDTO> create(DepartamentoDTO departamentoDTO) {
-
-        var optNome = repository.findByNome(departamentoDTO.getNome());
-        if (optNome.isPresent()) {
-            throw new IllegalArgumentException("Departamento já existe.");
+ 
+        if (nomeJaExiste(departamentoDTO.getNome())) {
+            throw new IllegalArgumentException("Já existe um departamento com esse nome.");
         }
 
-        List<FuncionarioDTO> funcionarios = funcionarioService.findAll();
-
-        if(!funcionarios.isEmpty() && departamentoDTO.getGerenteId() == null){
-            throw new CampoDisponivelVazioException("O campo gerente não pode ser vazio, " +
+        if(utils.entidadesFuncionarioExistem() && departamentoDTO.getGerenteId() == null){
+            throw new CampoVazioException("O campo gerente não pode ser vazio, " +
             "já existem funcionários disponíveis no sistema.");
         }
 
-        var gerenteId = departamentoDTO.getGerenteId();
-        if (gerenteId != null) {
-            var departamentoGerente = repository.findByGerenteId(gerenteId);
-    
-            if (departamentoGerente.isPresent() && !departamentoGerente.get().getId().equals(departamentoDTO.getId())) {
-                String gerenteNome = funcionarioService.findById(gerenteId).get().getNome();
-                String departamentoNome = departamentoGerente.get().getNome();
-                throw new GerenteJaAssociadoException(gerenteNome, departamentoNome);
-            }
+        var depGerenciado = utils.funcionarioGerenciaDeparamento(departamentoDTO.getGerenteId());
+
+        if (depGerenciado.isPresent()) {
+            throw new GerenteJaAssociadoException(depGerenciado.get().getGerente().getNome(),
+            depGerenciado.get().getNome()); 
         }
 
         var departamentoSalvo = repository.save(mapper.toDepartamento(departamentoDTO));
@@ -85,7 +71,6 @@ public class DepartamentoServiceImpl implements DepartamentoService {
         var departamentoSalvo = repository.findById(id);
 
         if (departamentoSalvo.isEmpty()){
-            System.out.println("Ele passou aquiiiiiiiiiiiiiiii");
             throw new EntityNotFoundException("Departamento não foi encontrado no sistema.");
         }
 
@@ -95,14 +80,14 @@ public class DepartamentoServiceImpl implements DepartamentoService {
     @Override
     public void remove(UUID id) {
         
-        var departamentoDTO = findById(id).get();
+        findById(id).get();
 
-        funcionarioService.findAll().forEach( (fun) -> {
-            if(fun.getDepartamento().getId() == departamentoDTO.getId()){
-                throw new RemocaoDepartamentoGerenteException("Não foi possível remover o departamento "
-                + departamentoDTO.getNome() + " pois o(a) gerente " + fun.getNome() + " está associado(a) à ele.");
-            }
-        });
+        var funcionarios = utils.funcionariosEmDepartamento(id);
+        
+        if(!funcionarios.isEmpty()){
+            throw new EntidadeAssociadaException("Não foi possível remover o departamento pois estes funcionários "
+                + funcionarios + " estão associados à ele.");
+        }
         
         repository.deleteById(id);
     }
@@ -112,21 +97,22 @@ public class DepartamentoServiceImpl implements DepartamentoService {
         
         findById(departamentoDTO.getId());
     
-        var gerenteId = departamentoDTO.getGerenteId();
-    
-        if (gerenteId != null) {
-            var departamentoGerente = repository.findByGerenteId(gerenteId);
-    
-            if (departamentoGerente.isPresent() && !departamentoGerente.get().getId().equals(departamentoDTO.getId())) {
-                String gerenteNome = funcionarioService.findById(gerenteId).get().getNome();
-                String departamentoNome = departamentoGerente.get().getNome();
-                throw new GerenteJaAssociadoException(gerenteNome, departamentoNome);
-            }
+        var depGerenciado = utils.funcionarioGerenciaDeparamento(departamentoDTO.getGerenteId());
+
+        if (depGerenciado.isPresent()) {
+            throw new GerenteJaAssociadoException(depGerenciado.get().getGerente().getNome(),
+            depGerenciado.get().getNome()); 
         }
 
         var departamentoSalvo = repository.save(mapper.toDepartamento(departamentoDTO));
 
         return Optional.of(mapper.toDepartamentoDTO(departamentoSalvo));
+    }
+
+    public Boolean nomeJaExiste(String nome){
+        var optCpf = repository.findByNome(nome);
+        if(optCpf.isPresent()) return true;
+        return false;
     }
     
 }
